fix(snapshot): refactor and isloate snapshot common api to independent file
