+++ io-engine/src/bin/io-engine-client/v1/replica_cli.rs
@@ -110,54 +110,6 @@ pub fn subcommands<'a, 'b>() -> App<'a, 'b> {
                 .index(1)
                 .help("Replica uuid"),
         );
-    let snapshot_create = SubCommand::with_name("snapshot_create")
-        .about("Replica snapshot")
-        .arg(
-            Arg::with_name("replica_uuid")
-                .required(true)
-                .index(1)
-                .help("Replica uuid"),
-        )
-        .arg(
-            Arg::with_name("snapshot_name")
-                .required(true)
-                .index(2)
-                .help("Snapshot name"),
-        )
-        .arg(
-            Arg::with_name("entity_id")
-                .required(true)
-                .index(3)
-                .help("Entity Id"),
-        )
-        .arg(
-            Arg::with_name("txn_id")
-                .required(true)
-                .index(4)
-                .help("Transaction id"),
-        )
-        .arg(
-            Arg::with_name("snapshot_uuid")
-                .required(true)
-                .index(5)
-                .help("Snapshot uuid"),
-        );
-    let snapshot_list = SubCommand::with_name("snapshot_list")
-        .about("List Replica Snapshot")
-        .arg(
-            Arg::with_name("replica_uuid")
-                .required(false)
-                .index(1)
-                .help("Replica uuid"),
-        );
-    let snapshot_delete = SubCommand::with_name("snapshot_delete")
-        .about("Delete Replica Snapshot")
-        .arg(
-            Arg::with_name("snapshot_uuid")
-                .required(true)
-                .index(1)
-                .help("Snapshot uuid"),
-        );
     SubCommand::with_name("replica")
         .settings(&[
             AppSettings::SubcommandRequiredElseHelp,
@@ -173,9 +125,6 @@ pub fn subcommands<'a, 'b>() -> App<'a, 'b> {
         .subcommand(
             SubCommand::with_name("stats").about("IO stats of replicas"),
         )
-        .subcommand(snapshot_create)
-        .subcommand(snapshot_list)
-        .subcommand(snapshot_delete)
 }
 
 pub async fn handler(
@@ -189,13 +138,6 @@ pub async fn handler(
         ("share", Some(args)) => replica_share(ctx, args).await,
         ("unshare", Some(args)) => replica_unshare(ctx, args).await,
         ("stats", Some(args)) => replica_stat(ctx, args).await,
-        ("snapshot_create", Some(args)) => {
-            replica_snapshot_create(ctx, args).await
-        }
-        ("snapshot_list", Some(args)) => replica_snapshot_list(ctx, args).await,
-        ("snapshot_delete", Some(args)) => {
-            replica_snapshot_delete(ctx, args).await
-        }
         (cmd, _) => {
             Err(Status::not_found(format!("command {cmd} does not exist")))
                 .context(GrpcStatus)
@@ -510,204 +452,6 @@ async fn replica_stat(
     Ok(())
 }
 
-/// Replica Snapshot Create CLI Function.
-async fn replica_snapshot_create(
-    mut ctx: Context,
-    matches: &ArgMatches<'_>,
-) -> crate::Result<()> {
-    let replica_uuid = matches
-        .value_of("replica_uuid")
-        .ok_or_else(|| ClientError::MissingValue {
-            field: "replica_uuid".to_string(),
-        })?
-        .to_owned();
-    let snapshot_name = matches
-        .value_of("snapshot_name")
-        .ok_or_else(|| ClientError::MissingValue {
-            field: "snapshot_name".to_string(),
-        })?
-        .to_owned();
-    let entity_id = matches
-        .value_of("entity_id")
-        .ok_or_else(|| ClientError::MissingValue {
-            field: "entity_id".to_string(),
-        })?
-        .to_owned();
-    let txn_id = matches
-        .value_of("txn_id")
-        .ok_or_else(|| ClientError::MissingValue {
-            field: "txn_id".to_string(),
-        })?
-        .to_owned();
-    let snapshot_uuid = matches
-        .value_of("snapshot_uuid")
-        .ok_or_else(|| ClientError::MissingValue {
-            field: "snapshot_uuid".to_string(),
-        })?
-        .to_owned();
-    // let snapshot_uuid = Uuid::generate().to_string();
-    let request = v1_rpc::replica::CreateReplicaSnapshotRequest {
-        replica_uuid,
-        snapshot_uuid,
-        snapshot_name,
-        entity_id,
-        txn_id,
-    };
-
-    let response = ctx
-        .v1
-        .replica
-        .create_replica_snapshot(request)
-        .await
-        .context(GrpcStatus)?;
-
-    match ctx.output {
-        OutputFormat::Json => {
-            println!(
-                "Snapshot Created {}",
-                serde_json::to_string_pretty(&response.get_ref())
-                    .unwrap()
-                    .to_colored_json_auto()
-                    .unwrap()
-            );
-        }
-        OutputFormat::Default => {
-            let snapshots = &response.get_ref().snapshot;
-            let table = snapshots
-                .iter()
-                .map(|r| {
-                    vec![
-                        r.snapshot_uuid.clone(),
-                        r.snapshot_name.clone(),
-                        r.snapshot_size.clone().to_string(),
-                        r.num_clones.to_string(),
-                        r.replica_uuid.clone(),
-                        r.replica_size.to_string(),
-                        r.entity_id.clone(),
-                        r.txn_id.clone(),
-                        r.valid_snapshot.to_string(),
-                    ]
-                })
-                .collect();
-            ctx.print_list(
-                vec![
-                    "SNAP_UUID",
-                    "SNAP_NAME",
-                    "SNAP_SIZE",
-                    "CLONE_COUNT",
-                    "REPLICA_UUID",
-                    "REPLICA_SIZE",
-                    "ENTITY_ID",
-                    "TXN_ID",
-                    "VALID_SNAPSHOT",
-                ],
-                table,
-            );
-        }
-    };
-
-    Ok(())
-}
-/// Replica Snapshot List CLI Function.
-async fn replica_snapshot_list(
-    mut ctx: Context,
-    matches: &ArgMatches<'_>,
-) -> crate::Result<()> {
-    let replica_uuid = matches.value_of("replica_uuid").map(|s| s.to_owned());
-    let request = v1_rpc::replica::ListReplicaSnapshotsRequest {
-        replica_uuid,
-    };
-
-    let response = ctx
-        .v1
-        .replica
-        .list_replica_snapshot(request)
-        .await
-        .context(GrpcStatus)?;
-    match ctx.output {
-        OutputFormat::Json => {
-            println!(
-                "{}",
-                serde_json::to_string_pretty(&response.get_ref())
-                    .unwrap()
-                    .to_colored_json_auto()
-                    .unwrap()
-            );
-        }
-        OutputFormat::Default => {
-            let snapshots = &response.get_ref().snapshots;
-            if snapshots.is_empty() {
-                ctx.v1("No snapshots found");
-                return Ok(());
-            }
-
-            let table = snapshots
-                .iter()
-                .map(|r| {
-                    vec![
-                        r.snapshot_uuid.clone(),
-                        r.snapshot_name.clone(),
-                        r.snapshot_size.to_string(),
-                        r.num_clones.to_string(),
-                        r.replica_uuid.clone(),
-                        r.replica_size.to_string(),
-                        r.entity_id.clone(),
-                        r.txn_id.clone(),
-                        r.valid_snapshot.to_string(),
-                    ]
-                })
-                .collect();
-            ctx.print_list(
-                vec![
-                    "SNAP_UUID",
-                    "SNAP_NAME",
-                    "SNAP_SIZE",
-                    "CLONE_COUNT",
-                    "REPLICA_UUID",
-                    "REPLICA_SIZE",
-                    "ENTITY_ID",
-                    "TXN_ID",
-                    "VALID_SNAPSHOT",
-                ],
-                table,
-            );
-        }
-    };
-
-    Ok(())
-}
-/// Replica Snapshot Delete CLI Function.
-async fn replica_snapshot_delete(
-    mut ctx: Context,
-    matches: &ArgMatches<'_>,
-) -> crate::Result<()> {
-    let snapshot_uuid = matches
-        .value_of("snapshot_uuid")
-        .ok_or_else(|| ClientError::MissingValue {
-            field: "snapshot_uuid".to_string(),
-        })?
-        .to_owned();
-
-    let _ = ctx
-        .v1
-        .replica
-        .delete_replica_snapshot(
-            v1_rpc::replica::DeleteReplicaSnapshotRequest {
-                snapshot_uuid: snapshot_uuid.clone(),
-            },
-        )
-        .await
-        .context(GrpcStatus)?;
-
-    match ctx.output {
-        OutputFormat::Json => {}
-        OutputFormat::Default => {
-            println!("snapshot: {} is deleted", &snapshot_uuid);
-        }
-    }
-
-    Ok(())
-}
 fn parse_replica_protocol(pcol: Option<&str>) -> Result<i32, Status> {
     match pcol {
         None => Ok(v1_rpc::common::ShareProtocol::None as i32),
