+++ io-engine/src/core/snapshot.rs
@@ -1,4 +1,4 @@
-use crate::{core::logical_volume::LogicalVolume, lvs::Lvol};
+use crate::lvs::Lvol;
 use async_trait::async_trait;
 use chrono::{DateTime, Utc};
 use serde::{Deserialize, Serialize};
@@ -35,39 +35,72 @@ impl SnapshotParams {
 /// Snapshot Descriptor to respond back as part of listsnapshot
 #[derive(Clone, Debug)]
 pub struct VolumeSnapshotDescriptor {
-    snapshot_uuid: String,
-    snapshot_size: u64,
-    num_clones: u64,
-    timestamp: DateTime<Utc>,
-    replica_uuid: String,
-    replica_size: u64,
-    snap_param: SnapshotParams,
+    pub snapshot_lvol: Lvol,
+    pub source_uuid: String,
+    pub source_size: u64,
+    pub snap_param: SnapshotParams,
+    pub num_clones: u64, /* TODO: Need to move to SnapshotParams part of
+                          * clone feature. */
+    pub timestamp: DateTime<Utc>, /* TODO: Need to move to SnapshotParams
+                                   * part of clone feature. */
     // set to false, if any of the snapshotdescriptor is not filled properly
-    valid_snapshot: bool,
+    pub valid_snapshot: bool,
 }
-
 impl VolumeSnapshotDescriptor {
     pub fn new(
-        snapshot: &Lvol,
+        snapshot_lvol: Lvol,
+        source_uuid: String,
+        source_size: u64,
+        snap_param: SnapshotParams,
         num_clones: u64,
         timestamp: DateTime<Utc>,
-        replica_uuid: String,
-        replica_size: u64,
-        snap_param: SnapshotParams,
         valid_snapshot: bool,
     ) -> Self {
         Self {
-            snapshot_uuid: snapshot.uuid(),
-            snapshot_size: snapshot.size(),
+            snapshot_lvol,
+            source_uuid,
+            source_size,
+            snap_param,
             num_clones,
             timestamp,
-            replica_uuid,
-            replica_size,
-            snap_param,
             valid_snapshot,
         }
     }
+    /// Get snapshot lvol.
+    pub fn snapshot_lvol(&self) -> &Lvol {
+        &self.snapshot_lvol
+    }
+    /// Get snapshot_uuid.
+    pub fn source_uuid(&self) -> String {
+        self.source_uuid.clone()
+    }
+
+    /// Give snapshot size.
+    pub fn source_size(&self) -> u64 {
+        self.source_size
+    }
+
+    /// Get SnapshotParameters.
+    pub fn snapshot_params(&self) -> &SnapshotParams {
+        &self.snap_param
+    }
+
+    /// Give number of clones.
+    pub fn num_clones(&self) -> u64 {
+        self.num_clones
+    }
+
+    /// Give timestamp of Snapshot creation.
+    pub fn timestamp(&self) -> DateTime<Utc> {
+        self.timestamp
+    }
+
+    /// Get ValidSnapshot value.
+    pub fn valid_snapshot(&self) -> bool {
+        self.valid_snapshot
+    }
 }
+
 /// Snapshot attributes used to store its properties.
 #[derive(Debug, EnumCountMacro, EnumIter)]
 pub enum SnapshotXattrs {
@@ -111,8 +144,12 @@ pub trait SnapshotOps {
         snap_uuid: &str,
     ) -> Option<SnapshotParams>;
 
-    /// List Snapshots.
-    fn list_snapshot(&self) -> Vec<VolumeSnapshotDescriptor>;
+    /// List Snapshot details based on source UUID from which snapshot is
+    /// created.
+    fn list_snapshot_by_source_uuid(&self) -> Vec<VolumeSnapshotDescriptor>;
+
+    /// List Single snapshot details based on snapshot UUID.
+    fn list_snapshot_by_snapshot_uuid(&self) -> Vec<VolumeSnapshotDescriptor>;
 }
 
 /// Traits gives the Snapshots Related Parameters.
@@ -146,33 +183,6 @@ pub trait SnapshotDescriptor {
 
     /// Set snapshot uuid of the snapshot.
     fn set_snapshot_uuid(&mut self, snapshot_uuid: String);
-}
-
-/// Traits gives VolumeSnapshot Descriptor.
-pub trait VolumeSnapshotDescriptors {
-    /// Get snapshot_uuid.
-    fn snapshot_uuid(&self) -> &str;
-
-    /// Get snapshot size.
-    fn snapshot_size(&self) -> u64;
-
-    /// Give number of clones.
-    fn num_clones(&self) -> u64;
-
-    /// Give timestamp of Snapshot creation.
-    fn timestamp(&self) -> DateTime<Utc>;
-
-    /// Give replica uuid.
-    fn replica_uuid(&self) -> &str;
-
-    /// Give replica size.
-    fn replica_size(&self) -> u64;
-
-    /// Get SnapshotParameters.
-    fn snapshot_params(&self) -> &SnapshotParams;
-
-    /// Get ValidSnapshot value.
-    fn valid_snapshot(&self) -> bool;
 }
 
 /// Trait to give interface for all Snapshot Parameters.
@@ -220,47 +230,5 @@ impl SnapshotDescriptor for SnapshotParams {
     /// Set snapshot uuid of the snapshot.
     fn set_snapshot_uuid(&mut self, snapshot_uuid: String) {
         self.snapshot_uuid = Some(snapshot_uuid);
-    }
-}
-
-impl VolumeSnapshotDescriptors for VolumeSnapshotDescriptor {
-    /// Get snapshot_uuid.
-    fn snapshot_uuid(&self) -> &str {
-        &self.snapshot_uuid
-    }
-
-    /// Give snapshot size.
-    fn snapshot_size(&self) -> u64 {
-        self.snapshot_size
-    }
-
-    /// Give number of clones.
-    fn num_clones(&self) -> u64 {
-        self.num_clones
-    }
-
-    /// Give timestamp of Snapshot creation.
-    fn timestamp(&self) -> DateTime<Utc> {
-        self.timestamp
-    }
-
-    /// Give replica uuid.
-    fn replica_uuid(&self) -> &str {
-        &self.replica_uuid
-    }
-
-    /// Give replica size.
-    fn replica_size(&self) -> u64 {
-        self.replica_size
-    }
-
-    /// Get SnapshotParameters.
-    fn snapshot_params(&self) -> &SnapshotParams {
-        &self.snap_param
-    }
-
-    /// Get ValidSnapshot value.
-    fn valid_snapshot(&self) -> bool {
-        self.valid_snapshot
     }
 }
