+++ test/python/README.md
@@ -62,6 +62,8 @@ python -m grpc_tools.protoc --proto_path=`realpath rpc/mayastor-api/protobuf/v1`
 python -m grpc_tools.protoc --proto_path=`realpath rpc/mayastor-api/protobuf/v1` --python_out=test/python --grpc_python_out=test/python replica.proto
 python -m grpc_tools.protoc --proto_path=`realpath rpc/mayastor-api/protobuf/v1` --python_out=test/python --grpc_python_out=test/python pool.proto
 python -m grpc_tools.protoc --proto_path=`realpath rpc/mayastor-api/protobuf/v1` --python_out=test/python --grpc_python_out=test/python nexus.proto
+python -m grpc_tools.protoc --proto_path=`realpath rpc/mayastor-api/protobuf/v1` --python_out=test/python --grpc_python_out=test/python snapshot.proto
+
 virtualenv --no-setuptools test/python/venv
 source test/python/venv/bin/activate
 pip install -r test/python/requirements.txt
