+++ test/python/setup.sh
@@ -12,7 +12,7 @@ cd "$SRCDIR"
 
 python -m grpc_tools.protoc --proto_path=rpc/mayastor-api/protobuf --grpc_python_out=test/python --python_out=test/python mayastor.proto
 python -m grpc_tools.protoc --proto_path=rpc/mayastor-api/protobuf/v1 --grpc_python_out=test/python --python_out=test/python \
-  bdev.proto common.proto nexus.proto pool.proto replica.proto host.proto registration.proto
+  bdev.proto common.proto nexus.proto pool.proto replica.proto host.proto registration.proto snapshot.proto
 
 virtualenv --no-setuptools test/python/venv
 (source ./test/python/venv/bin/activate && pip install -r test/python/requirements.txt)
